+++ platforms/wildfly/src/main/webapp/META-INF/services/io.hawt.system.AuthenticatorFactory
@@ -1,0 +1,1 @@
+io.hawt.wildfly.WildflyAuthenticatorFactory
